+++ setup.py
@@ -1,4 +1,5 @@
 import subprocess
+
 import setuptools
 
 # This will fail if something happens or if not in a git repository.
